# docs/swagger.yaml
basePath: /api
definitions:
  model.CreateCategoryRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  model.UpdateCategoryRequest:
    properties:
      id:
        type: integer
      name:
        type: string
    required:
    - id
    - name
    type: object
  model.WebResponse:
    properties:
      code:
        type: integer
      data: {}
      error:
        items: {}
        type: array
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Ikhlash Mulyanurahman
    url: https://www.ikhlashmula.github.io/
  description: This is a sample project following clean architecture.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: golang-clean-architecture-project-structure
  version: "1.0"
paths:
  /categories:
    get:
      consumes:
      - application/json
      description: Get category
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.WebResponse'
      summary: Get category
      tags:
      - category
    post:
      consumes:
      - application/json
      description: create new category
      parameters:
      - description: request body category
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.WebResponse'
      summary: create category
      tags:
      - category
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: delete category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.WebResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.WebResponse'
      summary: delete category
      tags:
      - category
    get:
      consumes:
      - application/json
      description: Get category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.WebResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.WebResponse'
      summary: Get category
      tags:
      - category
    put:
      consumes:
      - application/json
      description: update category
      parameters:
      - description: request body category
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateCategoryRequest'
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.WebResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.WebResponse'
      summary: update category
      tags:
      - category
swagger: "2.0"